pyproject.toml
[project]
name = "whisper-input"
version = "0.2.0"
description = "语音输入工具 - 按住快捷键说话，松开自动输入"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "pystray>=0.19.0",
    "pillow>=10.0.0",
    "funasr>=1.0.0",
    "torch==2.5.1",
    "torchaudio==2.5.1",
    # Linux 专用
    "evdev>=1.7.0; sys_platform == 'linux'",
    "pygobject>=3.56.2; sys_platform == 'linux'",
    # macOS 专用
    "pynput>=1.7.6; sys_platform == 'darwin'",
]

[[tool.uv.index]]
name = "tuna"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://mirror.sjtu.edu.cn/pytorch-wheels/cu121"
explicit = true

[tool.ruff]
line-length = 80

[tool.ruff.lint]
extend-select = [
    "I",      # isort - import 排序
    "N",      # pep8-naming - 命名规范
    "UP",     # pyupgrade - 使用新语法
    "B",      # flake8-bugbear - 常见 bug
    "SIM",    # flake8-simplify - 可简化的代码
    "RUF",    # ruff 专有规则
]
ignore = [
    "RUF001",  # 字符串中的全角字符（中文项目正常）
    "RUF002",  # docstring 中的全角字符
    "RUF003",  # 注释中的全角字符
]

# Linux 从 SJTU CUDA 镜像装 torch，macOS 从默认 PyPI 装
[tool.uv.sources]
torch = [
    { index = "pytorch-cu121", marker = "sys_platform == 'linux'" },
]
torchaudio = [
    { index = "pytorch-cu121", marker = "sys_platform == 'linux'" },
]

[dependency-groups]
dev = [
    "ruff>=0.15.9",
]
